Start updating to latest version of z88dk.
Build tools -> python3
Makefile -> macos-ish

This is very crude updating - not bothering at all with the
structure, just fixing up 12 years of bugfixes/drift.

Two outstanding issues:

1. Border colour flips to white initially
2. Jumping just jumps too high

Main gotchas:

- __z88dk_callee needs to be marked as __naked or the stack is
double adjusted
- Do the same to __z88dk_fastcall for consistency

Layout in memory for sp1 is slightly in the wrong place - it
ends up in contended memory at the moment, issues raised on z88dk
to resolve.

+++ tools/png2sprite.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 """
 png2sprite.py
 Copyright (C) 2014 by Juan J. Martinez - usebox.net
@@ -74,7 +74,7 @@ def main():
 
     tiles = []
     # this is wrong, but assuming nxn sprites works
-    for j in range(h / 8):
+    for j in range(int(h / 8)):
         if not args.nomask:
             tiles.extend([255, 0, 255, 0, 255, 0, 255, 0])
